+++ R/autodetec.R
@@ -6,7 +6,8 @@
 #'   power = 1, bp = NULL, osci = FALSE, wl = 512, xl = 1, picsize = 1, res = 100, 
 #'   flim = c(0,22), ls = FALSE, sxrow = 10, rows = 10, mindur = NULL, maxdur = 
 #'   NULL, redo = FALSE, img = TRUE, it = "jpeg", set = FALSE, flist = NULL, smadj = NULL,
-#'   parallel = 1, path = NULL, pb = TRUE)
+#'   parallel = 1, path = NULL, pb = TRUE, pal = reverse.gray.colors.2, 
+#'   fast.spec = FALSE, ...)
 #' @param X Data frame with results from \code{\link{manualoc}} function or any data frame with columns
 #' for sound file name (sound.files), selection number (selec), and start and end time of signal
 #' (start and end). 
@@ -77,6 +78,17 @@
 #' If \code{NULL} (default) then the current working directory is used.
 #' @param pb Logical argument to control progress bar. Default is \code{TRUE}. Note that progress bar is only used
 #' when parallel = 1.
+#' @param pal Color palette function for spectrogram. Default is reverse.gray.colors.2. See 
+#' \code{\link[seewave]{spectro}} for more palettes. Palettes as \code{\link[monitoR]{gray.2}} may work better when \code{fast.spec = TRUE}.
+#' @param fast.spec Logical. If \code{TRUE} then image function is used internally to create spectrograms, which substantially 
+#' increases performance (much faster), although some options become unavailable, as collevels, and sc (amplitude scale).
+#' This option is indicated for signals with high background noise levels. Palette colors \code{\link[monitoR]{gray.1}}, \code{\link[monitoR]{gray.2}}, 
+#' \code{\link[monitoR]{gray.3}}, \code{\link[monitoR]{topo.1}} and \code{\link[monitoR]{rainbow.1}} (which should be imported from the package monitoR) seem
+#' to work better with 'fast.spec' spectograms. Palette colors \code{\link[monitoR]{gray.1}}, \code{\link[monitoR]{gray.2}}, 
+#' \code{\link[monitoR]{gray.3}} offer 
+#' decreasing darkness levels. THIS IS STILL BEING TESTED.
+#' @param ... Additional arguments to be passed to a modified version of \code{\link[seewave]{spectro}} for customizing
+#' graphical output.
 #' @return Image files with spectrograms showing the start and end of the detected signals. It 
 #'   also returns a data frame containing the start and end of each signal by 
 #'   sound file and selection number.
@@ -122,7 +134,8 @@ autodetec<-function(X= NULL, threshold=15, envt="abs", ssmooth = NULL, msmooth =
                     bp = NULL, osci = FALSE, wl = 512, xl = 1, picsize = 1, res = 100, flim = c(0,22), 
                     ls = FALSE, sxrow = 10, rows = 10, mindur = NULL, maxdur = NULL, redo = FALSE, 
                     img = TRUE, it = "jpeg", set = FALSE, flist = NULL, smadj = NULL, parallel = 1, 
-                    path = NULL, pb = TRUE){
+                    path = NULL, pb = TRUE, pal = reverse.gray.colors.2,
+                    fast.spec = FALSE, ...){
   
   #check path to working directory
   if(!is.null(path))
@@ -338,10 +351,18 @@ autodetec<-function(X= NULL, threshold=15, envt="abs", ssmooth = NULL, msmooth =
     if(!is.null(mindur)) time.song <-time.song[time.song$duration > mindur,]
     if(!is.null(maxdur)) time.song <-time.song[time.song$duration < maxdur,]
     
-    if(nrow(time.song)>0) 
+    if(nrow(time.song) > 0) 
     {if(xprov) time.song$selec <- paste(X$selec[i], 1:nrow(time.song), sep = "-") else
       time.song$selec <- 1:nrow(time.song)}
+   
+    #if nothing was detected
+    if(nrow(time.song)==0)
+      time.song <- data.frame(sound.files = X$sound.files[i], duration = NA,selec = NA,start = NA, end = NA) 
     
+    time.song1 <- time.song
+    
+    time.song$start[is.na(time.song$start)] <- -2
+    time.song$start[is.na(time.song$start)] <- -1
     
     
     if(!ls & img & nrow(time.song) > 0) {
@@ -355,8 +376,8 @@ autodetec<-function(X= NULL, threshold=15, envt="abs", ssmooth = NULL, msmooth =
         imgfun(filename = paste(fna, paste0(".", it), sep = "-"), 
         width = (10.16) * xl * picsize, height = (10.16) * picsize, units = "cm", res = res)
         
-      seewave::spectro(song, f = f, wl = wl, collevels=seq(-45,0,1),grid = FALSE, main = as.character(X$sound.files[i]), osc = osci,
-              scale = FALSE, palette = seewave::reverse.gray.colors.2, flim = fl)
+     spectro.INTFUN(song, f = f, wl = wl, collevels=seq(-45,0,1),grid = FALSE, main = as.character(X$sound.files[i]), osc = osci,  colwave = "blue4", fast.spec = fast.spec,
+              scale = FALSE, palette = pal, flim = fl, ...)
       rm(song)
       if(nrow(time.song)>0)
       {sapply(1:nrow(time.song), function(j)  abline(v=c(time.song$start[j]-X$start[i], time.song$end[j]-X$start[i]),col="red",lwd=2, lty= "dashed"))
@@ -370,15 +391,13 @@ autodetec<-function(X= NULL, threshold=15, envt="abs", ssmooth = NULL, msmooth =
  dev.off()
  }  
   }
-    #if nothing was detected
-  if(nrow(time.song)==0)
-  time.song<-data.frame(sound.files = X$sound.files[i], duration = NA,selec = NA,start = NA, end = NA)
+ 
   
     #remove duration column
-  time.song <- time.song[,grep("duration",colnames(time.song),invert = TRUE)]
+  time.song1 <- time.song1[,grep("duration",colnames(time.song1), invert = TRUE)]
   
-  return(time.song)
-  on.exit(rm(time.song))
+  return(time.song1)
+  on.exit(rm(time.song1))
   }
   
 
@@ -446,13 +465,18 @@ autodetec<-function(X= NULL, threshold=15, envt="abs", ssmooth = NULL, msmooth =
   {if(smadj == "start" | smadj == "both") results$start <- results$start-((threshold*2.376025e-07)-1.215234e-05)*ssmooth 
   if(smadj == "end" | smadj == "both")  results$end <- results$end-((threshold*-2.369313e-07)+1.215129e-05)*ssmooth }  
 
+  results1 <- results
+  
+  #remove NAs so the ones with no detections are printed
+  results$start[is.na(results$start)] <- -2
+  results$end[is.na(results$end)] <- -1
   
   # long spectrograms
   if(ls & img) {  
    if(any(parallel == 1, Sys.info()[1] == "Linux") & pb) message("Producing long spectrogram:")
     
     #function for long spectrograms (based on lspec function)
-    lspeFUN2 <- function(X, z, fl = flim, sl = sxrow, li = rows, fli = fli, pal) {
+    lspeFUN2 <- function(X, z, fl = flim, sl = sxrow, li = rows, fli = fli, pal, fast.spec) {
     #subset for a sound file    
     Y <- X[!is.na(X$start) & X$sound.files == z, ]
       
@@ -491,8 +515,8 @@ autodetec<-function(X= NULL, threshold=15, envt="abs", ssmooth = NULL, msmooth =
         while(x <= li-1){
           x <- x + 1
           if(all(((x)*sl+li*(sl)*(j-1))-sl < dur & (x)*sl+li*(sl)*(j-1) < dur)){  #for rows with complete spectro
-            seewave::spectro(rec, f = f, wl = 512, flim = frli, tlim = c(((x)*sl+li*(sl)*(j-1))-sl, (x)*sl+li*(sl)*(j-1)), 
-                    ovlp = 10, collevels = collev, grid = gr, scale = FALSE, palette = pal, axisX = TRUE)
+            spectro.INTFUN(rec, f = f, wl = 512, flim = frli, tlim = c(((x)*sl+li*(sl)*(j-1))-sl, (x)*sl+li*(sl)*(j-1)), collevels = collev, grid = gr, scale = FALSE, palette = pal, axisX = TRUE,
+                    fast.spec = fast.spec, ...)
             if(x == 1)  text((sl-0.01*sl) + (li*sl)*(j - 1), frli[2] - (frli[2]-frli[1])/10, paste(substring(z, first = 1, 
        last = nchar(as.character(z))-4), "-p", j, sep = ""), pos = 2, font = 2, cex = cex)
           
@@ -504,10 +528,10 @@ autodetec<-function(X= NULL, threshold=15, envt="abs", ssmooth = NULL, msmooth =
             } else 
               { #for rows with incomplete spectro (final row)
   if(all(((x)*sl+li*(sl)*(j-1))-sl < dur & (x)*sl+li*(sl)*(j-1) > dur)){ 
-    seewave::spectro(seewave::pastew(seewave::noisew(f = f, d = (x)*sl+li*(sl)*(j-1)-dur+1, type = "unif",   
+    spectro.INTFUN(seewave::pastew(seewave::noisew(f = f, d = (x)*sl+li*(sl)*(j-1)-dur+1, type = "unif",   
     listen = FALSE,  output = "Wave"), seewave::cutw(wave = rec, f = f, from = ((x)*sl+li*(sl)*(j-1))-sl,
      to = dur, output = "Wave"), f =f,  output = "Wave"), f = f, wl = 512, flim = frli, 
-     tlim = c(0, sl), ovlp = 10, collevels = collev, grid = gr, scale = FALSE, palette = pal, axisX = FALSE)
+     tlim = c(0, sl), collevels = collev, grid = gr, scale = FALSE, palette = pal, axisX = FALSE, fast.spec = fast.spec, ...)
                                        
     if(x == 1)  text((sl-0.01*sl) + (li*sl)*(j - 1), frli[2] - (frli[2]-frli[1])/10, paste(substring(z, first = 1,                                                                                              last = nchar(as.character(z))-4), "-p", j, sep = ""), pos = 2, font = 2, cex = cex)
     
@@ -515,18 +539,31 @@ autodetec<-function(X= NULL, threshold=15, envt="abs", ssmooth = NULL, msmooth =
   axis(1, at = c(0:sl), labels = c((((x)*sl+li*(sl)*(j-1))-sl):((x)*sl+li*(sl)*(j-1))) , tick = TRUE)
         
   
-  if(nrow(Y)> 0)
+  
+  
+  if(nrow(Y) > 0)
   {
     abline(v = c(Y$start, Y$end) - (((x)*sl+li*(sl)*(j-1))-sl), col = "red", lty = 2)
     text(x = ((Y$start + Y$end)/2) - (((x)*sl+li*(sl)*(j-1))-sl),  frli[2] - 2*((frli[2] - frli[1])/12), labels = Y$selec, font = 4)
   }
   
   #add line indicating end of sound file
-  abline(v = dur-(((x)*sl+li*(sl)*(j-1))-sl), lwd = 2.5)} else 
+  abline(v = dur-(((x)*sl+li*(sl)*(j-1))-sl), lwd = 2.5)
+  usr<-par("usr")    
+  polygon(x = rep(c(sl - ((x)*sl+li*(sl)*(j-1)-dur), usr[2]), each = 2), y = c(usr[3], usr[4], usr[4], usr[3]), col = "white")
+  
+  #add text indicating end of sound files
+  text(dur-(((x)*sl+li*(sl)*(j-1))-sl), frli[2]-(frli[2]-frli[1])/2, "END OF SOUND FILE", pos = 4, font = 2, cex = 1.1)
+  
+  
+  } else 
     {
     plot(1, 1, col = "white", col.axis =  "white", col.lab  =  "white", 
     xaxt = "n", yaxt = "n")
     
+      usr<-par("usr")    
+      polygon(x = rep(c(sl - ((x)*sl+li*(sl)*(j-1)-dur), usr[2]), each = 2), y = c(usr[3], usr[4], usr[4], usr[3]), col = "white")
+      
     #add text indicating end of sound files
   text(dur-(((x)*sl+li*(sl)*(j-1))-sl), frli[2]-(frli[2]-frli[1])/2, "END OF SOUND FILE", pos = 4, font = 2, cex = 1.1)
     }
@@ -547,7 +584,7 @@ autodetec<-function(X= NULL, threshold=15, envt="abs", ssmooth = NULL, msmooth =
     
     a1 <- parallel::parLapply(cl, unique(results$sound.files), function(z)
     {
-      lspeFUN2(X = results, z = z, fl = flim, sl = sxrow, li = rows, pal = seewave::reverse.gray.colors.2)
+      lspeFUN2(X = results, z = z, fl = flim, sl = sxrow, li = rows, pal = pal, fast.spec = fast.spec)
     })
     
     parallel::stopCluster(cl)
@@ -558,11 +595,11 @@ autodetec<-function(X= NULL, threshold=15, envt="abs", ssmooth = NULL, msmooth =
       
       if(pb)
         a1 <- pbmcapply::pbmclapply(unique(results$sound.files), mc.cores = parallel, function(z) {
-          lspeFUN2(X = results, z = z, fl = flim, sl = sxrow, li = rows, pal = seewave::reverse.gray.colors.2)
+          lspeFUN2(X = results, z = z, fl = flim, sl = sxrow, li = rows, pal = pal, fast.spec = fast.spec)
         }) else
         
       a1 <- parallel::mclapply(unique(results$sound.files), mc.cores = parallel, function(z) {
-        lspeFUN2(X = results, z = z, fl = flim, sl = sxrow, li = rows, pal = seewave::reverse.gray.colors.2)
+        lspeFUN2(X = results, z = z, fl = flim, sl = sxrow, li = rows, pal = pal, fast.spec = fast.spec)
       })
     }
     if(!any(Sys.info()[1] == c("Linux", "Windows"))) # parallel in OSX
@@ -570,7 +607,7 @@ autodetec<-function(X= NULL, threshold=15, envt="abs", ssmooth = NULL, msmooth =
       cl <- parallel::makeForkCluster(getOption("cl.cores", parallel))
       
       sp <- foreach::foreach(i = unique(results$sound.files)) %dopar% {
-        lspeFUN2(X = results, z = z, fl = flim, sl = sxrow, li = rows, pal = seewave::reverse.gray.colors.2)
+        lspeFUN2(X = results, z = z, fl = flim, sl = sxrow, li = rows, pal = pal, fast.spec = fast.spec)
       }
       
       parallel::stopCluster(cl)
@@ -580,16 +617,16 @@ autodetec<-function(X= NULL, threshold=15, envt="abs", ssmooth = NULL, msmooth =
     if(pb)
      a1 <- pbapply::pblapply(unique(results$sound.files), function(z) 
   { 
-  lspeFUN2(X = results, z = z, fl = flim, sl = sxrow, li = rows, pal = seewave::reverse.gray.colors.2)
+  lspeFUN2(X = results, z = z, fl = flim, sl = sxrow, li = rows, pal = pal, fast.spec = fast.spec)
   }) else  a1 <- lapply(unique(results$sound.files), function(z) 
   { 
-    lspeFUN2(X = results, z = z, fl = flim, sl = sxrow, li = rows, pal = seewave::reverse.gray.colors.2)
+    lspeFUN2(X = results, z = z, fl = flim, sl = sxrow, li = rows, pal = pal, fast.spec = fast.spec)
   })
   
   }
   }
 
-return(results)
+return(results1)
   if(img) on.exit(dev.off())
   if(!is.null(path)) setwd(wd)
 }