+++ R/frange.detec.R
@@ -10,7 +10,7 @@
 #'   and producing the spectrogram (using \code{\link[seewave]{spectro}}, if \code{plot = TRUE}). 
 #' @param fsmooth A numeric vector of length 1 to smooth the frequency spectrum with a mean
 #'  sliding window in kHz. This help to average amplitude "hills" to minimize the effect of
-#'  amplitude modulation.
+#'  amplitude modulation. Default is 0.1.
 #' @param threshold Amplitude threshold (\%) for fundamental frequency and 
 #'   dominant frequency detection. Default is 10.
 #' @param wn Character vector of length 1 specifying window name. Default is 
@@ -60,135 +60,15 @@
 
 frange.detec <- function(wave, wl = 512, fsmooth = 0.1, threshold = 10, wn = "hanning", flim = c(0, 22), bp = NULL, fast.spec = FALSE, ovlp = 50, pal = reverse.gray.colors.2, widths = c(2, 1), min.range = NULL, main = NULL, plot = TRUE)
 {
-  # get sampling rate
-  f <-  wave@samp.rate
   
-  # fix flim
-  flim <- c(0, floor(f/2000))
-  if(flim[2] > ceiling(f/2000) - 1) flim[2] <- ceiling(f/2000) - 1 
+  frng <- frd.INTFUN(wave = wave, wl = wl, fsmooth = fsmooth, threshold = threshold, wn = wn, flim = flim, bp = bp, ovlp = ovlp, min.range = min.range)
   
-  
-  # mean spectrum
-  spc <- meanspec(wave, plot = F, wl = wl, f = f, wn = wn, ovlp = ovlp)
-  
-  # get frequency windows length for smoothing
-  step <- wave@samp.rate/wl/1000
-  
-  fsmooth <- fsmooth/step
-  
-  # number of samples
-  n <- nrow(spc)
-  
-  # smoothing parameter
-  FWL <- fsmooth - 1
-  
-  # smooth 
-  z <- apply(as.matrix(1:(n - FWL)), 1, function(y) sum(spc[y:(y + 
-                                                                 FWL), 2]))
-  zf <- seq(min(spc[,1]), max(spc[,1]), length.out = length(z))
-  
-  if(!is.null(bp)) { 
-    #remove range outsde bp
-    z <- z[zf > bp[1] & zf < bp[2]]
-    zf <- zf[zf > bp[1] & zf < bp[2]]
-    
-    # make minimum amplitude 0
-    z <- z - min(z)
-    z[z < 0] <- 0}
-  
-  # normalize amplitude from 0 to 1
-  z <- z/max(z)
-  z1 <- rep(0, length(z))
-  
-  #get freqs crossing threshold
-  z1[z > threshold/100] <- 1 
-  z2 <- z1[2:length(z1)] - z1[1:(length(z1) - 1)]
-  z2 <- c(0, z2)
-  
-  start <- zf[z2 == 1]
-  end <- zf[z2 == -1]
-  
-  # set low and hi f to flim/bp if not detected
-  if(!is.null(bp)) { 
-    min.start <- ifelse(length(start) == 0 || is.infinite(min(start)), yes = bp[1], no = min(start))
-    max.end <- ifelse(length(end) == 0 || is.infinite(min(end)), yes = bp[2], no = max(end))
-    
-    if(!is.null(min.range) && max.end - min.start < min.range) max.end <- bp[2]
-  } else {
-    min.start <- ifelse(length(start) == 0 || is.infinite(min(start)), yes = flim[1], no = min(start))
-    max.end <- ifelse(length(end) == 0 || is.infinite(min(end)), yes = flim[2], no = max(end))
-    
-    if(!is.null(min.range) && max.end - min.start < min.range) max.end <- flim[2]
-  }
   
   if(plot)
-  {
-    # split screen
-    m <- rbind(c(0, widths[1]/sum(widths), 0, 0.93), #1
-               c(widths[1]/sum(widths), 1, 0 , 0.93),
-               c(0, 1,  0.93 , 1)) #3 
-    
-    invisible(close.screen(all.screens = TRUE))  
-    split.screen(m)
-    screen(1)
-    par(mar = c(3.4, 3.4, 0.5, 0))
-    
-    # create spectro
-    spectro2(wave = wave, f = f, flim = flim, fast.spec = fast.spec, palette = pal, ovlp = ovlp, wl = wl, grid = F, tlab = "", flab = "")
-    
-    #add line highlighting freq range
-    abline(h = c(min.start, max.end), col = "#80C3FF", lty = 3, lwd = 3.3)
-    
-    # add axis labels
-    mtext(side = 1, text = "Time (s)", line = 2.3)
-    mtext(side = 2, text = "Frequency (kHz)", line = 2.3)
-    
-    #second plot
-    screen(2)
-    par(mar = c(3.4, 0, 0.5, 0.5))
-    
-    plot(z, zf, type = "l", ylim = flim, yaxs = "i", xaxs = "i", yaxt = "n", xlab = "", col = "white", xaxt = "n")
-    
-    # add axis& labels
-    axis(1, at = seq(0.2, 1, by = 0.4))
-    mtext(side = 1, text = "Amplitude (%)", line = 2.3)
-    
-    # fix amplitude values to close polygon (just for ploting)
-    z3 <- c(0, z, 0)
-    
-    if(!is.null(bp)) zf3 <- c(bp[1], zf, bp[2]) else zf3 <- c(flim[1], zf, flim[2])
-    
-    # plot amplitude values curve
-    polygon(cbind(z3, zf3), col= adjustcolor("#4D69FF", 0.9))
-    
-    # add border line
-    points(z3, zf3, type = "l", col = adjustcolor("black", 0.5))
-    
-    # add bacground color
-    rect(xleft = 0, ybottom = flim[1], xright = 1, ytop = flim[2], col = adjustcolor("#4D69FF", 0.05))
-    
-    # add gray boxes in filtered out freq bands
-    if(!is.null(bp))
-    {  rect(xleft = 0, ybottom = bp[2], xright = 1, ytop = flim[2], col = adjustcolor("gray", 0.5)) 
-      rect(xleft = 0, ybottom = flim[1], xright = 1, ytop = bp[1], col = adjustcolor("gray", 0.5))
-    }
-    
-    # add line to highligh freq range
-    abline(v = threshold/100, col = adjustcolor("blue4", 0.7), lty = 3, lwd = 2.3)
-    abline(h = c(min.start, max.end), col = "#80C3FF", lty = 3, lwd = 3.3)
-    
-    if(!is.null(main))
-    {
-      screen(3)
-      par( mar = rep(0, 4))
-      plot(0.5, xlim = c(0, 1), ylim = c(0, 1), type = "n", axes = FALSE, xlab = "", ylab = "", xaxt = "n", yaxt = "n")
-      
-      text(x = 0.5, y = 0.35, labels = main, font = 2)  
-    }
-    
-  }
-  # return low and high freq
-  return(data.frame(low.f = min.start, high.f = max.end))
+    frd.plot.INTFUN(wave = wave, min.start = frng$frange$low.f, max.end = frng$frange$high.f, af.mat = frng$af.mat, wl = wl, threshold = threshold, wn = wn, flim = flim, bp = bp, fast.spec = fast.spec, ovlp = ovlp, pal = pal, widths = widths, main = main)   
+
+    # return low and high freq
+  return(frng$frange)
   
   # close screens
   on.exit(invisible(close.screen(all.screens = TRUE)))