+++ R/spectro.INTFUN.2.R
@@ -1,7 +1,7 @@
 #internal warbleR function, not to be called by users. It is a modified version of seewave::spectro 
 # that allows to plot spectrograms without resetting the graphic device.It also allow using image() 
 # which substantially increases speed (although makes some options unavailable)
-spectro2 <- function(wave, f, wl = 512, wn = "hanning", zp = 0, ovlp = 0, 
+spectro.INTFUN.2 <- function(wave, f, wl = 512, wn = "hanning", zp = 0, ovlp = 0, 
                      complex = FALSE, norm = TRUE, fftw = FALSE, dB = "max0", 
                      dBref = NULL, plot = TRUE, grid = TRUE, 
                      cont = FALSE, collevels = NULL, palette = spectro.colors,